refactor(fw-tests): don't run fw tests that call t8n in parallel
