+++ src/ethereum_test_specs/tests/test_expect.py
@@ -57,6 +57,7 @@ def t8n() -> ExecutionSpecsTransitionTool:  # noqa: D103
 
 
 # Storage value mismatch tests
+@pytest.mark.run_in_serial
 @pytest.mark.parametrize(
     "pre,post,expected_exception",
     [
@@ -122,6 +123,7 @@ def test_post_storage_value_mismatch(
 
 
 # Nonce value mismatch tests
+@pytest.mark.run_in_serial
 @pytest.mark.parametrize(
     "pre,post",
     [
@@ -157,6 +159,7 @@ def test_post_nonce_value_mismatch(pre: Alloc, post: Alloc, state_test, t8n, for
 
 
 # Code value mismatch tests
+@pytest.mark.run_in_serial
 @pytest.mark.parametrize(
     "pre,post",
     [
@@ -192,6 +195,7 @@ def test_post_code_value_mismatch(pre: Alloc, post: Alloc, state_test, t8n, fork
 
 
 # Balance value mismatch tests
+@pytest.mark.run_in_serial
 @pytest.mark.parametrize(
     "pre,post",
     [
@@ -227,6 +231,7 @@ def test_post_balance_value_mismatch(pre: Alloc, post: Alloc, state_test, t8n, f
 
 
 # Account mismatch tests
+@pytest.mark.run_in_serial
 @pytest.mark.parametrize(
     "pre,post,exception_type",
     [
