+++ src/ethereum_test_tools/tests/test_code.py
@@ -294,6 +294,7 @@ def test_opcodes_if(conditional_bytecode: bytes, expected: bytes):
     assert bytes(conditional_bytecode) == expected
 
 
+@pytest.mark.run_in_serial
 @pytest.mark.parametrize(
     "tx_data,switch_bytecode,expected_storage",
     [
