+++ src/pytest_plugins/filler/tests/conftest.py
@@ -8,14 +8,6 @@ import sysconfig
 import pytest
 
 
-def pytest_runtest_setup(item):
-    """Hook to skip tests if running with pytest-xdist in parallel."""
-    marker = item.get_closest_marker(name="run_in_serial")
-    if marker is not None:
-        if os.getenv("PYTEST_XDIST_WORKER_COUNT") not in [None, "1"]:
-            pytest.skip("Skipping test because pytest-xdist is running with more than one worker.")
-
-
 @pytest.fixture(autouse=True)
 def monkeypatch_path_for_entry_points(monkeypatch):
     """
